requirements.txt
Django==4.2
django-admin-sortable2==2.2.6
django-tinymce==4.1.0
environs==14.1.1
pillow==11.1.0
